+++ src/main/java/com/minimallifestyleproject/intership/user/UserService.java
@@ -2,6 +2,10 @@ package com.minimallifestyleproject.intership.user;
 
 import com.minimallifestyleproject.intership.security.CustomUserDetails;
 import com.minimallifestyleproject.intership.security.JwtService;
+import com.minimallifestyleproject.intership.user.dto.LoginRequestDto;
+import com.minimallifestyleproject.intership.user.dto.LoginResponseDto;
+import com.minimallifestyleproject.intership.user.dto.UserRequestDto;
+import com.minimallifestyleproject.intership.user.dto.UserResponseDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.authentication.AuthenticationManager;
