+++ src/main/java/com/minimallifestyleproject/intership/user/dto/UserRequestDto.java
@@ -1,4 +1,4 @@
-package com.minimallifestyleproject.intership.user;
+package com.minimallifestyleproject.intership.user.dto;
 
 import lombok.Data;
